requirements.txt
groq
chromadb
fastapi
uvicorn[standard]
python-dotenv
neo4j>=5.14.0

# Auth
python-jose[cryptography]
passlib[bcrypt]
cryptography
httpx
email-validator

# Rate limiting
slowapi

# Async file response
aiofiles

# PDF report generation
reportlab>=4.0.0

# File upload support
python-multipart
pdfplumber
pypdf
